requirements.txt
fastapi
uvicorn
python-telegram-bot==20.6
python-dotenv
psycopg2-binary
requests
